requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0
python-dotenv>=1.0.0
numpy>=1.24.0
setuptools>=68.0.0
python-dotenv>=1.0.0